refactor: add unit tests for RecentlyViewedDao methods
